pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pathreview"
version = "0.1.0"
description = "AI-powered portfolio review assistant"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1,<5.0.0",
    "openai>=1.10.0",
    "chromadb>=0.4.22",
    "redis>=5.0.0",
    "httpx>=0.26.0",
    "pypdf>=3.17.0",
    "tiktoken>=0.5.2",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "pytest-httpserver>=1.0.8",
    "hypothesis>=6.92.0",
    "ruff>=0.2.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "mutmut>=2.4.4",
    "types-redis>=4.6.0",
]

[tool.setuptools.packages.find]
include = ["api*", "core*", "ingestion*", "rag*", "agent*", "safety*"]
exclude = ["frontend*", "alembic*", "tests*", "scripts*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"core/models/*.py" = ["E501"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (require Docker services)",
    "benchmark: Performance benchmarks",
    "security: Security and red-team tests",
]
